move resource search to it's own controller
also give a view to the Organization object type
