+++ app/views/stafftools/organizations/_organization.html.erb
@@ -1,0 +1,13 @@
+<div class="flex-table classroom-user">
+  <div class="flex-table-item">
+    <%= image_tag "#{organization.avatar_url}&size=96", height: 48, width: 48, class: 'avatar left', alt: "@#{organization.login}" %>
+
+    <div class="student-information">
+      <%= link_to organization.html_url do %>
+        <strong class="css-truncate-target login"><%= organization.login %></strong>
+      <% end %>
+    </div>
+
+    <span class="text-muted css-truncate css-truncate-target"><%= organization.name %></span>
+  </div>
+</div>
