+++ app/views/stafftools/resources/_search_results.html.erb
