--- spec/support/cassettes/StafftoolsController/GET_search/as_an_unauthorized_user/returns_a_404.json
@@ -1,1 +1,0 @@
-{"http_interactions":[{"request":{"method":"get","uri":"https://api.github.com/user","body":{"encoding":"US-ASCII","base64_string":""},"headers":{"Accept":["application/vnd.github.v3+json"],"User-Agent":["Octokit Ruby Gem 4.1.1"],"Content-Type":["application/json"],"Authorization":["token \u003cTEST_CLASSROOM_OWNER_GITHUB_TOKEN\u003e"],"Cache-Control":["no-cache, no-store"],"Accept-Encoding":["gzip;q=1.0,deflate;q=0.6,identity;q=0.3"]}},"response":{"status":{"code":200,"message":"OK"},"headers":{"Server":["GitHub.com"],"Date":["Tue, 17 Nov 2015 17:11:26 GMT"],"Content-Type":["application/json; charset=utf-8"],"Transfer-Encoding":["chunked"],"Status":["200 OK"],"X-Ratelimit-Limit":["5000"],"X-Ratelimit-Remaining":["4969"],"X-Ratelimit-Reset":["1447782428"],"Cache-Control":["private, max-age=60, s-maxage=60"],"Last-Modified":["Mon, 16 Nov 2015 20:28:56 GMT"],"Etag":["W/\"c2d1d0997fbc2420248fa5ec36f4b9ca\""],"X-Oauth-Scopes":["admin:org, delete_repo, repo, user:email"],"X-Accepted-Oauth-Scopes":[""],"X-Oauth-Client-Id":["\u003cTEST_APPLICATION_GITHUB_CLIENT_ID\u003e"],"Vary":["Accept, Authorization, Cookie, X-GitHub-OTP","Accept-Encoding"],"X-Github-Media-Type":["github.v3; format=json"],"X-Xss-Protection":["1; mode=block"],"X-Frame-Options":["deny"],"Content-Security-Policy":["default-src 'none'"],"Access-Control-Allow-Credentials":["true"],"Access-Control-Expose-Headers":["ETag, Link, X-GitHub-OTP, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset, X-OAuth-Scopes, X-Accepted-OAuth-Scopes, X-Poll-Interval"],"Access-Control-Allow-Origin":["*"],"Strict-Transport-Security":["max-age=31536000; includeSubdomains; preload"],"X-Content-Type-Options":["nosniff"],"X-Served-By":["e724c57ebb9961c772a91e2dd7421c8d"],"X-Github-Request-Id":["A46B0E3B:146C3:7DEE998:564B5FBE"]},"body":{"encoding":"ASCII-8BIT","base64_string":"eyJsb2dpbiI6InRhcmVieXRlIiwiaWQiOjxURVNUX0NMQVNTUk9PTV9PV05F\nUl9HSVRIVUJfSUQ+LCJhdmF0YXJfdXJsIjoiaHR0cHM6Ly9hdmF0YXJzLmdp\ndGh1YnVzZXJjb250ZW50LmNvbS91LzxURVNUX0NMQVNTUk9PTV9PV05FUl9H\nSVRIVUJfSUQ+P3Y9MyIsImdyYXZhdGFyX2lkIjoiIiwidXJsIjoiaHR0cHM6\nLy9hcGkuZ2l0aHViLmNvbS91c2Vycy90YXJlYnl0ZSIsImh0bWxfdXJsIjoi\naHR0cHM6Ly9naXRodWIuY29tL3RhcmVieXRlIiwiZm9sbG93ZXJzX3VybCI6\nImh0dHBzOi8vYXBpLmdpdGh1Yi5jb20vdXNlcnMvdGFyZWJ5dGUvZm9sbG93\nZXJzIiwiZm9sbG93aW5nX3VybCI6Imh0dHBzOi8vYXBpLmdpdGh1Yi5jb20v\ndXNlcnMvdGFyZWJ5dGUvZm9sbG93aW5ney9vdGhlcl91c2VyfSIsImdpc3Rz\nX3VybCI6Imh0dHBzOi8vYXBpLmdpdGh1Yi5jb20vdXNlcnMvdGFyZWJ5dGUv\nZ2lzdHN7L2dpc3RfaWR9Iiwic3RhcnJlZF91cmwiOiJodHRwczovL2FwaS5n\naXRodWIuY29tL3VzZXJzL3RhcmVieXRlL3N0YXJyZWR7L293bmVyfXsvcmVw\nb30iLCJzdWJzY3JpcHRpb25zX3VybCI6Imh0dHBzOi8vYXBpLmdpdGh1Yi5j\nb20vdXNlcnMvdGFyZWJ5dGUvc3Vic2NyaXB0aW9ucyIsIm9yZ2FuaXphdGlv\nbnNfdXJsIjoiaHR0cHM6Ly9hcGkuZ2l0aHViLmNvbS91c2Vycy90YXJlYnl0\nZS9vcmdzIiwicmVwb3NfdXJsIjoiaHR0cHM6Ly9hcGkuZ2l0aHViLmNvbS91\nc2Vycy90YXJlYnl0ZS9yZXBvcyIsImV2ZW50c191cmwiOiJodHRwczovL2Fw\naS5naXRodWIuY29tL3VzZXJzL3RhcmVieXRlL2V2ZW50c3svcHJpdmFjeX0i\nLCJyZWNlaXZlZF9ldmVudHNfdXJsIjoiaHR0cHM6Ly9hcGkuZ2l0aHViLmNv\nbS91c2Vycy90YXJlYnl0ZS9yZWNlaXZlZF9ldmVudHMiLCJ0eXBlIjoiVXNl\nciIsInNpdGVfYWRtaW4iOmZhbHNlLCJuYW1lIjoiTWFyayBUYXJlc2hhd3R5\nIiwiY29tcGFueSI6bnVsbCwiYmxvZyI6Imh0dHA6Ly9tYXJrdGFyZXNoYXd0\neS5jb20iLCJsb2NhdGlvbiI6IkNvbHVtYnVzLCBPaGlvIiwiZW1haWwiOm51\nbGwsImhpcmVhYmxlIjpudWxsLCJiaW8iOm51bGwsInB1YmxpY19yZXBvcyI6\nOTcsInB1YmxpY19naXN0cyI6MTIsImZvbGxvd2VycyI6NzgsImZvbGxvd2lu\nZyI6MTQ2LCJjcmVhdGVkX2F0IjoiMjAxMS0wMS0xNFQwNTozNzo1OFoiLCJ1\ncGRhdGVkX2F0IjoiMjAxNS0xMS0xNlQyMDoyODo1NloifQ==\n"},"http_version":null},"recorded_at":"Tue, 17 Nov 2015 17:11:26 GMT"}],"recorded_with":"VCR 2.9.3"}
